- Updated to latest coreplot code. - Enabled a number of important warnings and fixed those warnings. - Added VTB Direktbank to institutes. - AnimationHelper: when fading in an NSWindow make it key or not depending on parameter. - Category analysis:   - optimized plot data (use double arrays instead NSDecimalNumber, load in one step instead value-by-value)   - optimized selection graph (dynamically reduce number of data points to not more than what can be shown actually)   - main graph (restrict info annotation updates). - About panel now shows/hides animated. - ImportController: use the defined fields instead that found in a line to parse data. Ignore anything after the last defined line (makes it more error tolerant).

+++ client/Source/ChipTanWindowController.h
@@ -18,7 +18,7 @@
 	
 	char		*bitString;
 	int			frequency;
-	int			currentCode;
+	NSUInteger  currentCode;
 	int			clock;
 	NSUInteger	codeLen;
 	NSString	*tan;
